// fe/src/types/enums.ts
export enum Tab {
  News = 'News',
  Paper = 'Paper',
}
export enum UrlStatus {
  Idle = 'idle',
  Loading = 'loading',
  Valid = 'valid',
  Invalid = 'invalid',
  Unavailable = 'unavailable',
}
export enum ModalType {
  None = 'none',
  TryToExit = 'tryToExit',
  Loading = 'loading',
  InvalidURL = 'invalidURL',
  UnavailableURL = 'unavailableURL',
}